refactor: move cli print logic to a separate module
Summary:

The `cli/main.ts` file currently contains all the logic for printing to the console. This makes the file large and difficult to maintain. This change refactors the printing logic out of `cli/main.ts` and into a separate module. This will improve the separation of concerns and make the code easier to read and test.

Test Plan:

Run the preflight checks to ensure that the changes are valid.
